requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
DateTime==4.3
Flask==2.0.1
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyArango==1.3.5
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
zope.interface==5.4.0
